backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
python-multipart==0.0.20
pydantic==2.12.5
PyMuPDF==1.26.5
google-genai>=1.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
reportlab>=4.0
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.5
